Remove the carousel
Material Design doesn't "out of the box" support carousels, so I've
decided to, for the moment, simply remove the carousel altogether.
There are ways we could potentially bring it back later if people think
it's valuable; [the Material Design for Bootstrap package has a carousel
implementation](https://mdbootstrap.com/javascript/carousel/), for example.