Convert cards to Material Design
This sort of converts all the faculty cards over Material Design. It
isn't really right, though, because the photos are handled really
badly (unlike in Bootstrap, where they were handled very nicely).

I'm not sure how to deal with this, and we need to roll this lab out
today, so I'm committing this to this branch and moving on. It would be
nice to come back to it, though.